requirements.txt
absl-py==0.9.0
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
apache-beam==2.23.0
appdirs==1.4.4
argh==0.26.2
asgiref==3.3.1
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
auto-py-to-exe==2.7.8
autopep8==1.4.4
avro-python3==1.8.1
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
bottle==0.12.18
bottle-websocket==0.2.9
Bottleneck==1.3.2
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
clyent==1.2.2
cmake==3.18.2.post1
colorama==0.4.3
comtypes==1.1.7
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==5.0.1
contextlib2==0.6.0.post1
crayons==0.4.0
crcmod==1.7
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
distlib==0.3.0
distributed==2.11.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
dlib==19.21.1
dnspython==1.15.0
dnspython3==1.15.0
docopt==0.6.2
docutils==0.16
Eel==0.12.4
emoji==1.6.1
entrypoints==0.3
et-xmlfile==1.0.1
fastavro==0.23.6
fastcache==1.1.0
ffmpeg==1.4
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
flickrapi==2.4.0
fsspec==0.6.2
future==0.18.2
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
glob2==0.7
globals==0.0.42
gnuplotlib==0.33
google-auth==1.17.1
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.1.0
h5py==2.10.0
hdfs==2.5.8
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.17.4
humanize==3.2.0
hypothesis==5.5.4
idna==2.5
igramscraper==0.3.2
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
imutils==0.5.3
insta-scrape==2.1.2
instascrape==0.1
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mcstatus==4.0.0
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==2.0.0
more-itertools==8.2.0
mouse==0.7.1
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==0.6.1
mss==6.0.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numberify==0.0.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
numpysane==0.28
oauth2client==3.0.0
oauthlib==3.1.0
object-detection==0.1
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pbr==5.4.5
pefile==2019.4.18
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.10
pkginfo==1.5.0.1
playsound==1.2.2
pluggy==0.13.1
ply==3.11
plyer==1.4.3
pokemonNames==0.1.0
prettytable==2.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.12.2
psutil==5.6.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptt==0.3
py==1.8.1
py-librus-api==0.3.1
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycaw==20181226
PyCocoa==20.6.30
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pydot==1.4.1
pydub==0.24.1
pyflakes==2.1.1
pygame==1.9.6
PyGetWindow==0.0.8
pyglet==1.5.7
Pygments==2.5.2
PyInstaller==3.6
pylint==2.4.4
pymongo==3.11.0
PyMouse==1.0
PyMsgBox==1.0.8
PyNaCl==1.3.0
pynput==1.6.8
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.0
pypiwin32==223
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.15.7
PyScreeze==0.1.26
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-helper==0.0.14
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-slugify==3.0.2
python-stdnum==1.17
PyTweening==1.0.3
pytz==2020.5
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
readme-renderer==26.0
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rope==0.16.0
rsa==4.2
Rtree==0.9.3
ruamel_yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
SoundCard==0.4.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
sqlparse==0.4.1
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-metadata==0.22.2
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tf-slim==1.2.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.3
tox==3.15.2
tqdm==4.42.1
traitlets==4.3.3
twine==3.2.0
typing-extensions==3.7.4.2
ujson==1.35
unicodecsv==0.14.1
urllib3==1.21.1
validate-email==1.3
virtualenv==20.0.23
watchdog==0.10.2
wcwidth==0.1.8
webdriver-manager==3.3.0
webencodings==0.5.1
Werkzeug==1.0.0
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
win10toast==0.9
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
yarg==0.1.9
YOLO3==1.0
youtube-dl==2020.12.29
youtube-mp3-downloader==1.4
zict==1.0.0
zipp==2.2.0
